requirements.txt
faiss-cpu
streamlit
google-generativeai
python-dotenv
langchain
langchain_google_genai
langchain_community
langchain_core
gradio
selenium
pillow
sentence-transformers
numpy